sign out button & display that email is verified
-created sign out button on nav bar
-currently signed in users can see if email is verified(mostly for testing purposes)
-email verification
