+++ src/App.js
@@ -9,6 +9,7 @@ from 'react-router-dom';
 import Home from './pages/home';
 import Login from './pages/login';
 import SignUp from './pages/signup';
+import SignOut from './pages/signout';
 import AuthDetails from './components/authdetails';
 import Navbar from './components/navbar';
 import Footer from './components/footer';
@@ -22,6 +23,7 @@ function App() {
           <Route path="/" element={<Home/>} />
           <Route path="/login" element={<><Login /><AuthDetails /></>} />
           <Route path="/signup" element={<SignUp />} />
+          <Route path="/signout" element={<SignOut />} />
         </Routes>
       </BrowserRouter>
       <Footer/>
