+++ src/components/authdetails.js
@@ -1,4 +1,4 @@
-import { onAuthStateChanged, signOut } from "firebase/auth";
+import { onAuthStateChanged } from "firebase/auth";
 import React, { useEffect, useState } from "react";
 import { auth } from "../firebaseConfig"
 
@@ -16,24 +16,12 @@ const AuthDetails = () => {
 
         return () => {
             listen();
-        }
+        };
     }, []);
-
-    const userSignOut = () => {
-        signOut(auth)
-        .then(() => {
-            console.log('Sign out successful');
-        })
-        .catch((error) => {
-            console.log(error);
-            // ..
-        });
-    }
-
 
     return (
         <div>
-            {authUser ? <><p>{`Signed In as ${authUser.email}`}</p><button onClick={userSignOut}>Sign Out</button></> : <p>Signed Out</p>}
+            {authUser ? <p>{`Signed In as ${authUser.email}, Email Verified: ${auth.currentUser.emailVerified}`}</p> : <p>Signed Out</p>}
         </div>
     );
 };
